increase warning level & fix warnings

+++ src/CrazyflieUSB.cpp
@@ -56,7 +56,7 @@ void CrazyflieUSB::sendPacket(
     if (status != LIBUSB_SUCCESS) {
         throw std::runtime_error(libusb_error_name(status));
     }
-    if (length != transferred) {
+    if (length != (uint32_t)transferred) {
         std::stringstream sstr;
         sstr << "Did transfer " << transferred << " but " << length << " was requested!";
         throw std::runtime_error(sstr.str());
@@ -102,7 +102,7 @@ void CrazyflieUSB::sendPacketNoAck(
     if (status != LIBUSB_SUCCESS) {
         throw std::runtime_error(libusb_error_name(status));
     }
-    if (length != transferred) {
+    if (length != (uint32_t)transferred) {
         std::stringstream sstr;
         sstr << "Did transfer " << transferred << " but " << length << " was requested!";
         throw std::runtime_error(sstr.str());
